# swagger-spec.json
{"openapi":"3.0.0","paths":{"/":{"get":{"operationId":"AppController_appInfo","parameters":[],"responses":{"200":{"description":""}},"tags":["Root"]}},"/info":{"get":{"operationId":"AppController_appInfo","parameters":[],"responses":{"200":{"description":""}},"tags":["Root"]}},"/ping":{"get":{"operationId":"AppController_ping","parameters":[],"responses":{"200":{"description":"","content":{"application/json":{"schema":{"type":"string"}}}}},"tags":["Root"]}},"/user/register":{"post":{"operationId":"UserController_register","summary":"注册","parameters":[],"requestBody":{"required":true,"content":{"application/json":{"schema":{"$ref":"#/components/schemas/UserDto"}}}},"responses":{"201":{"description":""}},"tags":["User Routes"]}},"/user/login":{"post":{"operationId":"UserController_login","summary":"登录","parameters":[],"requestBody":{"required":true,"content":{"application/json":{"schema":{"$ref":"#/components/schemas/UserDto"}}}},"responses":{"200":{"description":""}},"tags":["User Routes"]}},"/user/admin":{"get":{"operationId":"UserController_adminInfo","summary":"获取管理员信息","parameters":[],"responses":{"200":{"description":"","content":{"application/json":{"schema":{"type":"object"}}}}},"tags":["User Routes"]}},"/user":{"get":{"operationId":"UserController_authorRank","summary":"随机获得指定数量的用户信息","parameters":[{"name":"size","required":true,"in":"path","schema":{"type":"number"}}],"responses":{"200":{"description":"","content":{"application/json":{"schema":{"type":"array","items":{"type":"object"}}}}}},"tags":["User Routes"]},"patch":{"operationId":"UserController_patchUserData","parameters":[],"requestBody":{"required":true,"content":{"application/json":{"schema":{"$ref":"#/components/schemas/UserDetailDto"}}}},"responses":{"200":{"description":"","content":{"application/json":{"schema":{"type":"object"}}}},"401":{"description":"Unauthorized"}},"tags":["User Routes"],"security":[{"bearer":[]}]},"put":{"operationId":"UserController_putUserData","parameters":[],"requestBody":{"required":true,"content":{"application/json":{"schema":{"$ref":"#/components/schemas/UserDetailDto"}}}},"responses":{"200":{"description":"","content":{"application/json":{"schema":{"type":"object"}}}},"401":{"description":"Unauthorized"}},"tags":["User Routes"],"security":[{"bearer":[]}]}},"/user/check_logged":{"get":{"operationId":"UserController_checkLogged","summary":"判断当前 Token 是否有效 ","parameters":[],"responses":{"200":{"description":""},"401":{"description":"Unauthorized"}},"tags":["User Routes"],"security":[{"bearer":[]}]}},"/user/{username}":{"get":{"operationId":"UserController_getUserInfo","summary":"获取指定用户名的信息","parameters":[{"name":"username","required":true,"in":"path","schema":{"type":"string"}}],"responses":{"200":{"description":""}},"tags":["User Routes"]}},"/auth/check_logged":{"get":{"operationId":"AuthController_checkLogged","summary":"判断当前 Token 是否有效 ","parameters":[],"responses":{"200":{"description":"","content":{"application/json":{"schema":{"type":"string"}}}},"401":{"description":"Unauthorized"}},"tags":["Auth Routes"],"security":[{"bearer":[]}]}},"/post":{"post":{"operationId":"PostController_create","parameters":[],"requestBody":{"required":true,"content":{"application/json":{"schema":{"$ref":"#/components/schemas/PostDto"}}}},"responses":{"201":{"description":"","content":{"application/json":{"schema":{"type":"object"}}}},"401":{"description":"Unauthorized"}},"tags":["Post Routes"],"security":[{"bearer":[]}]},"get":{"operationId":"PostController_getPaginate","summary":"分页获取博文","parameters":[{"name":"tag","required":false,"in":"query","schema":{"type":"string"}},{"name":"sort","required":false,"in":"query","schema":{"enum":["newest","three_days_hottest","weekly_hottest","monthly_hottest","hottest"],"type":"string"}},{"name":"pageCurrent","required":true,"in":"query","schema":{"type":"number"}},{"name":"pageSize","required":true,"in":"query","schema":{"type":"number"}}],"responses":{"200":{"description":""}},"tags":["Post Routes"]}},"/post/{id}":{"get":{"operationId":"PostController_findLostById","summary":"根据 id 查文章","parameters":[{"name":"id","required":true,"in":"path","schema":{"type":"string"}}],"responses":{"200":{"description":"","content":{"application/json":{"schema":{"type":"object"}}}}},"tags":["Post Routes"]},"delete":{"operationId":"PostController_deletePost","parameters":[{"name":"id","required":true,"in":"path","schema":{"type":"string"}}],"responses":{"200":{"description":""},"401":{"description":"Unauthorized"}},"tags":["Post Routes"],"security":[{"bearer":[]}]},"patch":{"operationId":"PostController_patch","parameters":[{"name":"id","required":true,"in":"path","example":"5e6f67e75b303781d2807278","schema":{"type":"string"}}],"requestBody":{"required":true,"content":{"application/json":{"schema":{"$ref":"#/components/schemas/PartialPostModel"}}}},"responses":{"200":{"description":""},"401":{"description":"Unauthorized"}},"tags":["Post Routes"],"security":[{"bearer":[]}]},"put":{"operationId":"PostController_put","parameters":[{"name":"id","required":true,"in":"path","example":"5e6f67e75b303781d2807278","schema":{"type":"string"}}],"requestBody":{"required":true,"content":{"application/json":{"schema":{"$ref":"#/components/schemas/PartialPostModel"}}}},"responses":{"200":{"description":""},"401":{"description":"Unauthorized"}},"tags":["Post Routes"],"security":[{"bearer":[]}]}}},"info":{"title":"API","description":"API description api/v1","version":"1","contact":{}},"tags":[],"servers":[],"components":{"securitySchemes":{"bearer":{"scheme":"bearer","bearerFormat":"JWT","type":"http"}},"schemas":{"UserDto":{"type":"object","properties":{"username":{"type":"string"},"password":{"type":"string"},"introduce":{"type":"string"},"avatar":{"type":"string"}},"required":["username","password","introduce","avatar"]},"UserDetailDto":{"type":"object","properties":{"username":{"type":"string"},"password":{"type":"string"},"introduce":{"type":"string"},"avatar":{"type":"string"},"admin":{"type":"boolean"},"_id":{"type":"string"}},"required":["username","password","introduce","avatar","_id"]},"PostDto":{"type":"object","properties":{"title":{"type":"string"},"content":{"type":"string"},"cover":{"type":"string"},"tags":{"type":"array","items":{"type":"string"}},"ad":{"type":"boolean"}},"required":["title","content","ad"]},"UserModel":{"type":"object","properties":{"username":{"type":"string"},"password":{"type":"string"},"authCode":{"type":"string"},"avatar":{"type":"string"},"introduce":{"type":"string"},"admin":{"type":"boolean"}},"required":["username","password","authCode","avatar","introduce","admin"]},"PartialPostModel":{"type":"object","properties":{"title":{"type":"string"},"content":{"type":"string"},"cover":{"type":"string"},"tags":{"type":"array","items":{"type":"string"}},"ad":{"type":"boolean"},"read":{"type":"number"},"user":{"$ref":"#/components/schemas/UserModel"}}}}}}